renamed user_input to constant.py.  moved params to constant.py
